(PC-1469) Moves class TestClient
