+++ pytest.ini
@@ -1,7 +1,7 @@
 [pytest]
-addopts=-v --capture=no
+addopts=-v
 testpaths=tests
 norecursedirs=.git venv/ .pytest_cache/
 python_files=*_test.py
-python_classes=*
+python_classes=Post Patch Put Get Delete Returns* *Test
 python_functions=test_* when_*
