+++ tests/routes/mediations_test.py
@@ -89,7 +89,7 @@ def test_create_mediation_with_thumb_file(app):
 
     auth_request = req_with_auth(email=user.email, password='p@55sw0rd')
 
-    with open(Path(path.dirname(path.realpath(__file__))) / '..'
+    with open(Path(path.dirname(path.realpath(__file__))) / '..' / '..'
               / 'sandboxes' / 'thumbs' / 'mediations' / 'FranckLepage', 'rb') as thumb_file:
         data = {
             'offerId': humanize(offer.id),
